Continues PR #36 with 50 more commits: per-model GPU/CPU hybrid tensor
placement (replacing the global GPU-crash latch), GGUF parser bug fixes,
and CLI/dependency repairs across 15 crates.

+++ Omnisystem/src/crates/universal-language-layer/src/ffi.rs
@@ -153,7 +153,7 @@ pub mod conversions {
         // This is a simplified version - real implementation would handle all types
         match &value.data {
             serde_json::Value::Null => Ok(std::ptr::null()),
-            serde_json::Value::Bool(b) => Ok(*b as *const libc::c_void),
+            serde_json::Value::Bool(b) => Ok((*b as i64) as *const libc::c_void),
             serde_json::Value::Number(n) => {
                 if let Some(i) = n.as_i64() {
                     Ok(i as *const libc::c_void)
